requirements.txt
torch
torchvision
timm
numpy
scikit-learn
Pillow